towards minimized/explicit groups: Avoid use of nobody/nogroup
This first came to my attention when I was looking at ntp-service-type
and realized that despite creating a dedicated system user account
("ntpd") and creating file(s) (in "/var/run/ntpd") the user was
assigned to the "nogroup" group. This was surprising to me as I thought
that typically no file should be owned by nobody/nogroup if it could be
anticipated.

After some googling I found that this was at least in agreement with the
Debian interpretation: "Daemons that need not own any files run as user
nobody and group nogroup. Thus, no files on a system should be owned by
this user or group."

Additionally, dedicated users/groups for daemons and the like are best
practice. Wikipedia summarizes nicely: "It was once common to run
daemons as nobody, especially on servers, in order to limit the damage
that could be done by a malicious user who gained control of
them. However, the usefulness of this technique is reduced if more than
one daemon is run like this, because then gaining control of one daemon
would provide control of them all. The reason is that processes owned by
the same user have the ability to send signals to each other and use
debugging facilities to read or even modify each other's memory. Modern
practice, as recommended by the Linux Standard Base, is to create a
separate user account for each daemon."

Change-Id: Ic44f7031a4eddaae467209f797bfa232af6b3456