gl upgrades: bump mesa-utils
And add a ton more deps/functionality.

Question about that install phase: there are a /lot/ more programs in
this package but only a handful are seleted, is that intentional (keep
the package small or whatever)? If so putting that phase back is simple
but there are probably some new programs (like maybe 'vkgears') to
consider including in the selection.

Change-Id: I7301a5512a52c7e8d9f5e4632397a2c590371592