fontconfig: use lastest git version
Releases are infrequent and (at the time of this writing) the newer
commits on master are few, useful, and seem safe. Fedora backports a
ton of them. Let's accomplish the same thing more easily and just use
the latest from master. Sticking to offiial releases isn't always the
right move, don't be scared!

Change-Id: I9e24c66a138ab71c0214d1277784b84570f6fe34