apple silicon gnulib test fixes
findutils/coreutils/m4 should ideally be rewritten to use gexps to the
need to use native-inputs for patches (which is fine but messier than
doing it the gexp way).

Note that the only commencement.scm package that needs the fix is
findutils-boot0. Other commencement packages inherit it
harmlessly (except for coreutils-boot0 where it's explicitly
deleted. The phase will fail without the patch as an input: another good
reason to use the gexp version in coreutils, in which case this deletion
can be removed).

Change-Id: I4a97d7f973734ec51c00eddca9abaf2121565642